+++ packages/plugin-build/src/config/tenon-react.config.ts
@@ -1,12 +1,14 @@
-import { Configuration,SourceMapDevToolPlugin } from 'webpack'
+import { Configuration,SourceMapDevToolPlugin,DefinePlugin } from 'webpack'
 import {TenonStylePlugin} from '@hummer/tenon-style-loader'
 import JsccPlugin from 'webpack-plugin-jscc'
 import {ProjectConfig} from '@hummer/cli-utils'
 import {getAssetsAddress} from '../utils/server'
 import * as path from 'path'
+import { BuildPlugin } from '..'
 
-export default function getTenonReactConfiguration(isProduction: boolean, hmConfig?:ProjectConfig): Configuration {
+export default function getTenonReactConfiguration(isProduction: boolean, hmConfig:ProjectConfig, context: BuildPlugin): Configuration {
   let plugins:any = []
+  let { map: needMap } = context.options
   if(hmConfig){
     if(hmConfig.jscc){
       plugins.push(new JsccPlugin(hmConfig.jscc))
@@ -21,9 +23,12 @@ export default function getTenonReactConfiguration(isProduction: boolean, hmConf
       append: '\n//# sourceMappingURL='+ getAssetsAddress() + '[url]'
     }))
   }
-  return {
-    mode: isProduction?'production':'development',
+  let devToolConfig = needMap ? {
     devtool: isProduction ? 'hidden-source-map' : 'cheap-module-source-map',
+  }: null;
+  return {
+    ...devToolConfig,
+    mode: isProduction?'production':'development',
     output: {
       publicPath: './'
     },
@@ -93,6 +98,8 @@ export default function getTenonReactConfiguration(isProduction: boolean, hmConf
     },
     plugins: [new TenonStylePlugin({
       packageName: '@hummer/tenon-react'
+    }),new DefinePlugin({
+      "NODE_DEBUG": JSON.stringify(isProduction? false: true)  // 控制是否注入 DevTool
     }), ...plugins]
   }
 }