+++ packages/cli/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [0.2.0](https://github.com/OrangeLab/hummer-cli/compare/v0.1.27...v0.2.0) (2021-06-18)
+## [0.3.1](https://github.com/OrangeLab/hummer-cli/compare/v0.3.0...v0.3.1) (2021-07-06)
 
 **Note:** Version bump only for package @hummer/cli
 
@@ -11,3 +11,14 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+# [0.3.0](https://github.com/OrangeLab/hummer-cli/compare/v0.2.0...v0.3.0) (2021-06-28)
+
+**Note:** Version bump only for package @hummer/cli
+
+
+
+
+
+# [0.2.0](https://github.com/OrangeLab/hummer-cli/compare/v0.1.27...v0.2.0) (2021-06-18)
+
+**Note:** Version bump only for package @hummer/cli