+++ packages/plugin-build/CHANGELOG.md
@@ -3,15 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1](https://github.com/OrangeLab/hummer-cli/compare/v0.3.0...v0.3.1) (2021-07-06)
+
+**Note:** Version bump only for package @hummer/cli-plugin-build
+
+
+
+
+
+# [0.3.0](https://github.com/OrangeLab/hummer-cli/compare/v0.2.0...v0.3.0) (2021-06-28)
+
+
+### Features
+
+* **build:** 打包时默认不生成 .map 文件(issue[#32](https://github.com/OrangeLab/hummer-cli/issues/32)) ([f5d16db](https://github.com/OrangeLab/hummer-cli/commit/f5d16db1a50fa7e8b3f59ca11c75c3b026d38a1b))
+
+
+
+
+
 # [0.2.0](https://github.com/OrangeLab/hummer-cli/compare/v0.1.27...v0.2.0) (2021-06-18)
 
 
 ### Features
 
 * tenon vue project support ts ([6e4235b](https://github.com/OrangeLab/hummer-cli/commit/6e4235bdc5f0fececdf7246b66a4289ff5217c13))
-
-
-
-
-
-