+++ packages/plugin-init/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [0.2.0](https://github.com/OrangeLab/hummer-cli/compare/v0.1.27...v0.2.0) (2021-06-18)
+# [0.3.0](https://github.com/OrangeLab/hummer-cli/compare/v0.2.0...v0.3.0) (2021-06-28)
 
 **Note:** Version bump only for package @hummer/cli-plugin-init
 
@@ -11,4 +11,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+# [0.2.0](https://github.com/OrangeLab/hummer-cli/compare/v0.1.27...v0.2.0) (2021-06-18)
 
+**Note:** Version bump only for package @hummer/cli-plugin-init